Add a type to encode the LL(1) parsing table
LL(1) is probably easier to implement than LR(1), so let's start by
going that direction. This change is a _tiny_ step in that direction by
adding an intensive and profound class abstracting the mind boggling
concept of the LL(1) parsing table into the most elegant of terms:
ll1_table.

In all seriousness, this class is absolutely tiny, but that's okay. It's
primary purpose is to help get me back into working on this project, but
it also has the added benefit of being a small added benefit.

The tests are woefully incomplete, but such is the nature of working on
a project for _fun_ in my spare time. It would be lovely if it were
tested more, but alas, my priority is to actually enjoy what I'm doing,
seeing as this is not my job :-).

I might enjoy writing flowery commit messages a bit too much ;-).